Improved Code Readability of API routes

+++ apps/backend/routers/userops.py
@@ -1,5 +1,10 @@
+"""
+userops.py
+This module contains the user operations routes for liking, bookmarking, subscribing to sources, updating profile, and managing user history.
+"""
+
 import logging
-from fastapi import APIRouter, HTTPException, Depends, status, Response
+from fastapi import APIRouter, HTTPException, Depends, Response
 from sqlalchemy.orm import Session
 from database.session import get_db
 from database.operations import get_user_history
@@ -16,13 +21,20 @@ router = APIRouter()
 
 @router.get("/userhistory")
 async def get_current_user_history(current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Returns the History of current history."""
+    """Returns the History of current history.
+
+    Args:
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        list: List of user history."""
     try:
         if current_user:
             history = get_user_history(current_user.id, db)
             return history
         raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED, detail="User not logged in.")
+            status_code=401, detail="User not logged in.")
     except Exception as e:
         logger.error(f"Error in retrieving current user history: {e}")
 
@@ -33,7 +45,15 @@ class ArticleRequest(BaseModel):
 
 @router.post("/like")
 async def like_article(request: ArticleRequest, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Like an article for a specific user"""
+    """Like an article for a specific user.
+
+    Args:
+        request (ArticleRequest): Article ID.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        bool: True if liked, False otherwise."""
     article_id = request.article_id
     if current_user:
         like = db.query(UserLikes).filter(
@@ -42,7 +62,8 @@ async def like_article(request: ArticleRequest, current_user: Users = Depends(ge
             db.delete(like)
         else:  # Like article
             like = UserLikes(user_id=current_user.id, article_id=article_id,
-                             liked_at=datetime.now(timezone.utc))  # Use timezone.utc
+                             # Use timezone.utc
+                             liked_at=datetime.now(timezone.utc))
             db.add(like)
         db.commit()
         return True
@@ -51,7 +72,14 @@ async def like_article(request: ArticleRequest, current_user: Users = Depends(ge
 
 @router.post("/bookmark")
 async def bookmark_article(request: ArticleRequest, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Bookmark an article for a specific user"""
+    """Bookmark an article for a specific user.
+
+    Args:
+        request (ArticleRequest): Article ID.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+    Returns:
+        bool: True if bookmarked, False otherwise."""
     article_id = request.article_id
     if current_user:
         bookmark = db.query(UserBookmarks).filter(
@@ -60,7 +88,8 @@ async def bookmark_article(request: ArticleRequest, current_user: Users = Depend
             db.delete(bookmark)
         else:
             bookmark = UserBookmarks(user_id=current_user.id, article_id=article_id,
-                                     bookmarked_at=datetime.now(timezone.utc))  # Use timezone.utc
+                                     # Use timezone.utc
+                                     bookmarked_at=datetime.now(timezone.utc))
             db.add(bookmark)
         db.commit()
         return True
@@ -73,20 +102,30 @@ class SubscriptionsRequest(BaseModel):
 
 @router.post("/subscribe")
 async def add_subscriptions(request: SubscriptionsRequest, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
+    """Subscribe or unsubscribe to a source.
+
+    Args:
+        request (SubscriptionsRequest): Source name.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: Subscription status."""
     try:
         source = request.source
-        source_id = db.query(Sources.id).filter(Sources.source == source).first()
-        
+        source_id = db.query(Sources.id).filter(
+            Sources.source == source).first()
+
         # If Source exists extract id else None
         source_id = source_id[0] if source_id else None
         if not source_id:
             raise HTTPException(status_code=404, detail="Source not found")
-        
+
         subscription = db.query(UserSubscriptions).filter(
             UserSubscriptions.user_id == current_user.id,
             UserSubscriptions.source_id == source_id
         ).first()
-        
+
         if subscription:  # Unsubscribe
             db.delete(subscription)
             db.commit()
@@ -102,6 +141,7 @@ async def add_subscriptions(request: SubscriptionsRequest, current_user: Users =
         raise e  # Re-raise HTTP exceptions
     except Exception as e:
         db.rollback()  # Rollback in case of any error
+        logger.error(f"Error in subscribing to source: {e}")
         raise HTTPException(
             status_code=500, detail=f"An error occurred: {str(e)}"
         )
@@ -109,9 +149,18 @@ async def add_subscriptions(request: SubscriptionsRequest, current_user: Users =
 
 @router.get("/getuser")
 async def get_user(current_user: Users = Depends(get_current_active_user)):
+    """Get current user details.
+
+    Args:
+        current_user (Users): Current active user.
+
+    Returns:
+        dict: Current user details."""
     if current_user:
         return {"username": current_user.username, "fullname": current_user.fullname, "email": current_user.email}
-    return None
+    logger.error("User not logged in")
+    raise HTTPException(
+        status_code=401, detail="User not logged in.")
 
 
 class UpdateProfile(BaseModel):
@@ -121,7 +170,15 @@ class UpdateProfile(BaseModel):
 
 @router.post("/updateprofile")
 async def update_profile(request: UpdateProfile, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Update User Profile"""
+    """Update User Profile.
+
+    Args:
+        request (UpdateProfile): User profile details.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: Profile update status."""
     user = db.query(Users).filter(Users.id == current_user.id).first()
 
     if not user:
@@ -134,6 +191,7 @@ async def update_profile(request: UpdateProfile, current_user: Users = Depends(g
         db.refresh(user)  # Refresh user instance with updated values
         return {"message": "Profile updated successfully"}
     except Exception as e:
+        logger.error(f"Error in updating profile: {e}")
         db.rollback()  # Rollback changes in case of failure
         raise HTTPException(
             status_code=500, detail=f"Error updating profile: {str(e)}")
@@ -141,11 +199,28 @@ async def update_profile(request: UpdateProfile, current_user: Users = Depends(g
 
 @router.get("/deleteaccount")
 async def delete_user(response: Response, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    user = db.query(Users).filter(Users.id == current_user.id).first()
-    db.delete(user)
-    db.commit()
-    response.delete_cookie("refresh_token")
-    return {"message": "User Deleted"}
+    """Delete User Account.
+
+    Args:
+        response (Response): Response object.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: Account deletion status."""
+    try:
+        user = db.query(Users).filter(Users.id == current_user.id).first()
+        if not user:
+            raise HTTPException(status_code=404, detail="User not found")
+        db.delete(user)
+        db.commit()
+        response.delete_cookie("refresh_token")
+        return {"message": "User Deleted"}
+    except Exception as e:
+        logger.error(f"Error in deleting user: {e}")
+        db.rollback()
+        raise HTTPException(
+            status_code=500, detail=f"Error deleting user: {str(e)}")
 
 
 class UpdatePassword(BaseModel):
@@ -155,6 +230,15 @@ class UpdatePassword(BaseModel):
 
 @router.post("/updatepassword")
 async def update_password(request: UpdatePassword, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
+    """Update User Password.
+
+    Args:
+        request (UpdatePassword): Password details.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: Password update status."""
     user = db.query(Users).filter(Users.id == current_user.id).first()
     if verify_password(request.oldPassword, user.hashed_password):
         try:
@@ -172,7 +256,14 @@ async def update_password(request: UpdatePassword, current_user: Users = Depends
 
 @router.get("/clearhistory")
 async def clear_history(current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Clear User History"""
+    """Clear User History.
+
+    Args:
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: History clearance status."""
     try:
         db.query(UserHistory).filter(
             UserHistory.user_id == current_user.id).delete()
@@ -186,7 +277,15 @@ async def clear_history(current_user: Users = Depends(get_current_active_user), 
 
 @router.get("/delete-history-item")
 async def delete_history_item(id: int, current_user: Users = Depends(get_current_active_user), db: Session = Depends(get_db)):
-    """Delete single history item"""
+    """Delete single history item.
+
+    Args:
+        id (int): Article ID.
+        current_user (Users): Current active user.
+        db (Session): Database session.
+
+    Returns:
+        dict: Deletion status."""
     try:
         item = db.query(UserHistory).filter(UserHistory.user_id ==
                                             current_user.id, UserHistory.article_id == id).first()
